refactor: 컨트롤러를 콘솔 인터페이스와 컨트롤러로 분리
- 웹으로 변경되어도 컨트롤러를 재사용할 수 있게 하기 위해
