+++ voucher-manager/src/main/java/com/wonu606/vouchermanager/VoucherManagerCommandLineRunner.java
@@ -1,6 +1,6 @@
 package com.wonu606.vouchermanager;
 
-import com.wonu606.vouchermanager.controller.VoucherController;
+import com.wonu606.vouchermanager.controller.VoucherConsoleInterface;
 import lombok.AllArgsConstructor;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 @AllArgsConstructor
 public class VoucherManagerCommandLineRunner implements CommandLineRunner {
 
-    private final VoucherController controller;
+    private final VoucherConsoleInterface controller;
 
     @Override
     public void run(String... args) {
